requirements-django.txt
python-dotenv
Django==5.1.1
djangorestframework==3.15.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
drf-yasg==1.21.7
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
certifi==2025.1.31
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
django-debug-toolbar==5.2.0
django-ratelimit==4.1.0
django-redis==5.4.0
fastapi==0.116.1
fonttools==4.59.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.9
kombu==5.4.2
matplotlib==3.10.5
multidict==6.2.0
nltk==3.9.1
numpy==2.3.2
openai==1.68.2
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.50
propcache==0.3.0
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
ratelimit==2.2.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.47.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.5
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
